a. Sigmund Bot Completed (Mastodon)/2. Error, GitHub doesn't run continually (works in Ubuntu Folder)/requirements.txt
mastodon.py
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
requests
